// app/types.tsx
export interface ButtonTypes {
  text: string;
  customFunction: () => {};
  isLoading: boolean;
}
export interface Match {
  fixtureID: string;
  teams: {
    home: { name: string; logo?: string };
    away: { name: string; logo?: string };
  };
  fixture: {
    date: string;
  };
  odds: { name: string; values: { odd: string }[] }[];
}

export interface UserContextTypes {
  user: User | null;
  register: (name: string, email: string, password: string) => Promise<void>;
  login: (email: string, password: string) => Promise<void>;
  logout: () => Promise<void>;
  checkUser: () => Promise<void>;
  gideon: { currency: string; balance: number };
}

export interface GameOddTypes {
  match: Match | null;
  date: string;
  sportType: string;
}

export interface User {
  id: string;
  name: string;
  bio: string;
  image: string;
  balance: number;
  status: boolean;
  rank: string;
}

export interface Profile {
  id: string;
  user_id: string;
  name: string;
  image: string;
  bio: string;
}

export interface RandomUsers {
  id: string;
  name: string;
  image: string;
}

export interface CropperDimensions {
  height?: number | null;
  width?: number | null;
  left?: number | null;
  top?: number | null;
}

export interface Like {
  id: string;
  user_id: string;
  post_id: string;
}

export interface Post {
  id: string;
  user_id: string;
  video_url: string;
  text: string;
  created_at: string;
}

export interface PostWithProfile {
  id: string;
  user_id: string;
  video_url: string;
  text: string;
  created_at: string;
  profile: {
    user_id: string;
    name: string;
    image: string;
  };
}

export interface CommentWithProfile {
  id: string;
  user_id: string;
  post_id: string;
  text: string;
  created_at: string;
  profile: {
    user_id: string;
    name: string;
    image: string;
  };
}

export interface Comment {
  id: string;
  user_id: string;
  post_id: string;
  text: string;
  created_at: string;
}

export interface ShowErrorObject {
  type: string;
  message: string;
}

export interface UploadError {
  type: string;
  message: string;
}

//////////////////////////////////////////////
//////////////////////////////////////////////

// COMPONENT TYPES
export interface CommentsHeaderCompTypes {
  params: { userId: string; postId: string };
  post: PostWithProfile;
}

export interface CommentsCompTypes {
  params: { userId: string; postId: string };
}

export interface PostPageTypes {
  params: { userId: string; postId: string };
}

export interface ProfilePageTypes {
  params: { id: string };
}

export interface SingleCommentCompTypes {
  params: { userId: string; postId: string };
  comment: CommentWithProfile;
}

export interface PostUserCompTypes {
  post: Post;
}

export interface PostMainCompTypes {
  post: PostWithProfile;
}

export interface PostMainLikesCompTypes {
  post: PostWithProfile;
}

export interface TextInputCompTypes {
  string: string;
  inputType: string;
  placeholder: string;
  onUpdate: (newValue: string) => void;
  error: string;
  label?: string;
  isDisabled?: boolean;
}

//////////////////////////////////////////////
//////////////////////////////////////////////

// LAYOUT INCLUDE TYPES
export interface MenuItemTypes {
  iconString: string;
  colorString: string;
  sizeString: string;
}

export interface NavItemTypes {
  text: string;
  link: string;
  icon: string;
}

export interface MenuItemFollowCompTypes {
  user: RandomUsers;
}